vcl/requirements/core.txt
awscli==1.22.73
Django==3.2.13
boto3==1.21.3
Celery==5.2.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-environ==0.8.1
psycopg2==2.8.6
django-redis==5.2.0
django-celery-beat==2.2.1
kubernetes==21.7.0
django-celery-results==2.2.0
drf-yasg==1.20.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
-e git+https://github.com/strata-kk/pylti1.3.git@f6969820f60431c03e7480452f5003aabfe44365#egg=pylti1p3
pika==1.2.0
whitenoise==6.0.0
gunicorn==20.1.0
gevent==21.12.0
